Update index.html
Updates

Main projects page:
Make the add a project circle bigger, move it so that it alings with the start timer icon in the cards.
Place Latest, Name, and Time, in a flat gray box and add a label to the left “Sort by:” Change Latest to Recent and change Time to Timer Amount.
Move Today: <time> over so that the right side of the text aligns with the right side of the timer icon.
Change the start button to a stop watch icon the same size as the current box the play button is in.


Issues: When I select the timer icon on the project details page for manual entry it won’t allow me to select cancel and close it. 
Move the pencil next to the title with some space and add below the description a button labled “Add Time”. Remove the stop watch icon.
Change Status to Project Status
Change “Tasks for this Project” to “Task List’.
Change the date selector to look like the single row bar that’s on the reports page. Move the Total Time for this Period box to above the date selector.

On the reports page when you select Custom for date it should present a custom date picker.Please fix this.  Right not it doesn’t do anything. This same date picker look and feel and functionality needs to be added to the project detail page.
We don’t show the time on any other page so remove the time from next to the day and date on this page

Accomplishment need to be editable and we should have the same date picker that we have on the project details page and the reports page. The form to enter an accomplishment can include a date accomplished on.

Let’s start a theme that all functional button are the coral color or some variation depending on the object.
Plus button is coral. The title of the page is color. This should be the same on all page for the title and the plus icon if the page has a plus icon. Change the day and date under each title to a light gray.
Sort box on the project page is a lighter version of coral.
The stop watch timer start on the project cards can remain gray.
Icons in the nav bar are coral and the background of the nav bar is slightly darker than the rest of the page (light mode) slightly lighter (dark mode)
In the New Project card New Project should be in coral. Remove the Add Details and Goals selector and just show the Time estimates box. 
On the Settings page move Dark Mode to the left of the selector to reduce the height of the card. 
Import from JSON and Export to JSON can be side by side and not one on top of the other to reduce this card height.

I’m adding a sample JSON export from a previous version of this app and I need this version to be able to import if. If data is missing that’s ok but I need it to import the project titles and recorded time so that a person using a previous version of this app can also use this version.